Add new scripts and update styles

+++ src/main/java/es/innovatech/DAO/ICategoriasDAO.java
@@ -1,20 +1,37 @@
 /*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
+ * Haz clic en nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt para cambiar esta licencia
+ * Haz clic en nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java para editar esta plantilla
  */
 package es.innovatech.DAO;
 
 import es.innovatech.beans.Categoria;
 import java.util.List;
+
 /**
- *
+ * La interfaz ICategoriasDAO define los métodos que deben ser implementados
+ * por las clases que manejan la interacción con la base de datos para obtener información relacionada con las categorías.
+ * 
  * @author pedro
  */
 public interface ICategoriasDAO {
-    
+
+    /**
+     * Obtiene la lista de categorías.
+     * 
+     * @return Lista de objetos Categoria
+     */
     public List<Categoria> getCategorias();
 
-    public Categoria getCategoria(int id);
+    /**
+     * Obtiene una categoría por su ID.
+     * 
+     * @param id ID de la categoría
+     * @return Objeto Categoria
+     */
+    public Categoria getCategoria(byte id);
 
+    /**
+     * Cierra la conexión con la base de datos.
+     */
     public void closeConnection();
 }
